third_party/blink/web_tests/editing/selection/stay-in-textarea-expected.txt

Before

After

Ensure that extending a selection beyond a textarea does not escape outside its shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textareaSelection.focusOffset is 2
PASS textareaSelection.focusNode is initialTextareaFocusNode
PASS textareaSelection.focusOffset is 3
PASS textareaSelection.focusNode is initialTextareaFocusNode
PASS textareaSelection.focusOffset is 4
PASS textareaSelection.focusNode is initialTextareaFocusNode
PASS textareaSelection.focusOffset is 4
PASS textareaSelection.focusNode is initialTextareaFocusNode
PASS textareaSelection.focusOffset is 4
PASS textareaSelection.focusNode is initialTextareaFocusNode
PASS textareaSelection.focusOffset is 4
PASS window.getSelection().focusNode is initialFocusNode
PASS window.getSelection().focusOffset is initialFocusOffset
PASS successfullyParsed is true

TEST COMPLETE
